dedup ansi color rendering.
