+++ src/vs/workbench/contrib/notebook/browser/view/output/transforms/textHelper.ts
@@ -12,8 +12,9 @@ import { Range } from 'vs/editor/common/core/range';
 import { PieceTreeTextBufferBuilder } from 'vs/editor/common/model/pieceTreeTextBuffer/pieceTreeTextBufferBuilder';
 import { IOpenerService } from 'vs/platform/opener/common/opener';
 import { IThemeService } from 'vs/platform/theme/common/themeService';
-import { handleANSIOutput } from 'vs/workbench/contrib/notebook/browser/view/output/transforms/errorTransform';
 import { ITextFileService } from 'vs/workbench/services/textfile/common/textfiles';
+import { handleANSIOutput } from 'vs/workbench/contrib/debug/browser/debugANSIHandling';
+import { LinkDetector } from 'vs/workbench/contrib/debug/browser/linkDetector';
 
 const SIZE_LIMIT = 65535;
 const LINES_LIMIT = 500;
@@ -49,7 +50,7 @@ function generateViewMoreElement(outputs: string[], openerService: IOpenerServic
 	return element;
 }
 
-export function truncatedArrayOfString(container: HTMLElement, outputs: string[], openerService: IOpenerService, textFileService: ITextFileService, themeService: IThemeService) {
+export function truncatedArrayOfString(container: HTMLElement, outputs: string[], linkDetector: LinkDetector, openerService: IOpenerService, textFileService: ITextFileService, themeService: IThemeService) {
 	const fullLen = outputs.reduce((p, c) => {
 		return p + c.length;
 	}, 0);
@@ -65,7 +66,7 @@ export function truncatedArrayOfString(container: HTMLElement, outputs: string[]
 		const sizeBufferLimitPosition = buffer.getPositionAt(SIZE_LIMIT);
 		if (sizeBufferLimitPosition.lineNumber < LINES_LIMIT) {
 			const truncatedText = buffer.getValueInRange(new Range(1, 1, sizeBufferLimitPosition.lineNumber, sizeBufferLimitPosition.column), EndOfLinePreference.TextDefined);
-			container.appendChild(handleANSIOutput(truncatedText, themeService));
+			container.appendChild(handleANSIOutput(truncatedText, linkDetector, themeService, undefined));
 			// view more ...
 			container.appendChild(generateViewMoreElement(outputs, openerService, textFileService));
 			return;
@@ -82,13 +83,13 @@ export function truncatedArrayOfString(container: HTMLElement, outputs: string[]
 	if (buffer.getLineCount() < LINES_LIMIT) {
 		const lineCount = buffer.getLineCount();
 		const fullRange = new Range(1, 1, lineCount, Math.max(1, buffer.getLineLastNonWhitespaceColumn(lineCount)));
-		container.appendChild(handleANSIOutput(buffer.getValueInRange(fullRange, EndOfLinePreference.TextDefined), themeService));
+		container.appendChild(handleANSIOutput(buffer.getValueInRange(fullRange, EndOfLinePreference.TextDefined), linkDetector, themeService, undefined));
 		return;
 	}
 
 	const pre = DOM.$('pre');
 	container.appendChild(pre);
-	pre.appendChild(handleANSIOutput(buffer.getValueInRange(new Range(1, 1, LINES_LIMIT - 5, buffer.getLineLastNonWhitespaceColumn(LINES_LIMIT - 5)), EndOfLinePreference.TextDefined), themeService));
+	pre.appendChild(handleANSIOutput(buffer.getValueInRange(new Range(1, 1, LINES_LIMIT - 5, buffer.getLineLastNonWhitespaceColumn(LINES_LIMIT - 5)), EndOfLinePreference.TextDefined), linkDetector, themeService, undefined));
 
 	// view more ...
 	container.appendChild(generateViewMoreElement(outputs, openerService, textFileService));
@@ -96,5 +97,5 @@ export function truncatedArrayOfString(container: HTMLElement, outputs: string[]
 	const lineCount = buffer.getLineCount();
 	const pre2 = DOM.$('div');
 	container.appendChild(pre2);
-	pre2.appendChild(handleANSIOutput(buffer.getValueInRange(new Range(lineCount - 5, 1, lineCount, buffer.getLineLastNonWhitespaceColumn(lineCount)), EndOfLinePreference.TextDefined), themeService));
+	pre2.appendChild(handleANSIOutput(buffer.getValueInRange(new Range(lineCount - 5, 1, lineCount, buffer.getLineLastNonWhitespaceColumn(lineCount)), EndOfLinePreference.TextDefined), linkDetector, themeService, undefined));
 }
